+++ packages/strapi-generate-new/lib/utils/db-questions.js
