+++ packages/strapi-generate-new/lib/utils/has-yarn.js
@@ -1,0 +1,13 @@
+'use strict';
+
+const execa = require('execa');
+
+module.exports = function hasYarn() {
+  try {
+    const { code } = execa.shellSync('yarnpkg --version');
+    if (code === 0) return true;
+    return false;
+  } catch (err) {
+    return false;
+  }
+};
