+++ packages/strapi/bin/strapi.js
@@ -80,6 +80,7 @@ program
 // `$ strapi new`
 program
   .command('new <directory>')
+  .option('--no-run', 'Disable quickstart app start after creation')
   .option('--debug', 'Display database connection error')
   .option('--quickstart', 'Quickstart app creation')
   .option('--dbclient <dbclient>', 'Database client')
